refactor: helmraiser -> helm
Renames the helmraiser package to helm, because its scope is going to be
extended in the following.

It will not only cover helmraiser functionality going on, but also house
declarative chart management code
