+++ pkg/helm/helm_test.go
@@ -1,4 +1,4 @@
-package helmraiser
+package helm
 
 import (
 	"fmt"
